requirements.txt
asgiref==3.3.1
bandit==1.7.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.3
django-ajax-datatable==4.1.4
dparse==0.5.1
gitdb==4.0.5
GitPython==3.1.11
gunicorn==20.0.4
idna==2.10
packaging==20.8
pbr==5.5.1
Pillow==8.0.1
psycopg2==2.8.6
pyparsing==2.4.7
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
safety==1.9.0
selenium==3.141.0
six==1.15.0
smmap==3.0.4
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
